feat: add writer type using deref and demut
Following this apporach to using defer and defmut to keep compability between the writer type and pgpool type.

This way we can guarantee all of these benefits:

Transparency: Allows Reader and Writer to be used as PgPool without modifying the code that expects a PgPool.

Encapsulation: Still maintains the ability to add methods or additional logic to the Reader and Writer structures in the future.

Flexibility: We can control which methods or properties to expose by implementing or not certain traits.